Add files via upload
Color detection is the process of detecting the name of any color. Simple isn’t it? Well, for humans this is an extremely easy task but for computers, it is not straightforward. 
In this color detection project, we are going to build an application through which you can automatically get the name of the color by clicking on them.